requirements.txt
django
djangorestframework
# WSGI server for production
gunicorn
